buy-tracer-web/requirements.txt
# Web 框架
Flask==3.0.0
Flask-CORS==4.0.0

# 數據處理
pandas==2.1.4
numpy==1.26.2

# 股票數據源
twstock==1.3.1

# 圖表生成
plotly==5.18.0

# 環境變數管理
python-dotenv==1.0.0

# WSGI 服務器
gunicorn==21.2.0
waitress==2.1.2

# 工具套件
python-dateutil==2.8.2
pytz==2023.3

# 速率限制（可選）
Flask-Limiter==3.5.0

# JSON 加速（可選）
ujson==5.9.0

# Redis 快取（可選）
# redis==5.0.1
# flask-caching==2.1.0

# 開發工具（開發環境使用）
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
isort==5.13.2

# 型別檢查（可選）
# mypy==1.8.0
